Increase number of reader threads on COS side
On COS side, all threads block on the pipe input when they start
but since we start threads for the blocks but write the tx and
tx-by-addr first, we end up in a deadlock because the tx threads
cannot be schedule because the pool is full with block threads
waiting to read. So allow 3 times as many threads for the COS read
side.